requirements.txt
asn1crypto==0.24.0
attrdict==2.0.1
attrs==17.4.0
Automat==0.6.0
base58==1.0.3
blinker==1.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cytoolz==0.10.0
dominate==2.3.5
eth-abi==2.0.0a1
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-tester==0.2.0b2
eth-utils==1.6.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
hexbytes==0.2.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
incremental==16.10.1
ipfshttpclient==0.4.12
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
lru-dict==1.1.6
MarkupSafe==1.1.1
multiaddr==0.0.8
netaddr==0.7.19
netifaces==0.10.4
oauthlib==2.0.6
parsimonious==0.8.1
py-solc-x==1.1.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pycryptodome==3.8.2
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-debian==0.1.32
pyxdg==0.25
PyYAML==3.12
requests==2.22.0
requests-unixsocket==0.1.5
rlp==1.1.0
SecretStorage==2.3.1
semantic-version==2.6.0
service-identity==16.0.0
simplejson==3.13.2
six==1.12.0
ssh-import-id==5.7
toolz==0.10.0
Twisted==17.9.0
urllib3==1.25.3
varint==1.0.2
virtualenv==16.6.2
visitor==0.1.3
web3==4.9.2
websockets==6.0
Werkzeug==0.15.5
WTForms==2.2.1
